refactor:Role Actions now RoleEventReactions.cs
It looks better and makes it all more clearly.
