+++ Program.cs
@@ -17,6 +17,7 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using DiscordBot.Rolemanager;
 using System.Collections.Specialized;
+using System.Threading.Channels;
 using Newtonsoft.Json.Linq;
 
 
@@ -28,16 +29,6 @@ namespace DiscordBot
     {
       var db = new Database();
       db.defaultSetup();
-      /*var tmp = db.runSQL("SELECT rolename, keycode FROM CustomRoles");
-
-      for (var i = 0; i < tmp.Count; i++)
-      {
-        for (var j = 0; j < tmp[i].Length; j++)
-        {
-          Console.WriteLine($"{j}" + tmp[i][j]);
-        }
-      }*/
-      
       //MainAsync().GetAwaiter().GetResult();
       GuildmemberTask().GetAwaiter().GetResult();
     }
@@ -94,10 +85,7 @@ namespace DiscordBot
         TokenType = TokenType.Bot,
         Intents = DiscordIntents.All
       });
-      DiscordChannel[] arrayChannels;
-      DiscordGuild[] arrayGuilds;
-      DiscordRole[] arrayRoles;
-      GuildMemberAddEventArgs globalMember = null;
+      RoleEventReactions RoleEvents = new RoleEventReactions();
       /*
        * Command Prefix you need, to use the Command.
        */
@@ -111,56 +99,14 @@ namespace DiscordBot
        */
       discord.GuildMemberAdded += async (s, e) =>
       {
-        await e.Member.CreateDmChannelAsync();
-        await e.Member.SendMessageAsync("Do you have any special Keycode?");
-        globalMember = e;
+        await RoleEvents.ReactOnUserJoin(s, e);
       };
       /*
        * Event that reacts on Message written from User.
        */
       discord.MessageCreated += async (s, e) =>
       {
-        var db = new Database();
-        var tmp = db.runSQL("SELECT rolename, keycode FROM CustomRoles");
-        /*
-         * Bot reacts to private User Message to get his defined Role.
-         * SQL Query template used for tests(Roles need to exist on Discord Server):
-         * create table CustomRoles
-          (
-            RoleID int primary key auto_increment,
-            rolename varchar(255),
-            keycode varchar(255)
-           );
-            insert into CustomRoles (rolename, keycode)
-            values("Moderator", "Moderator123");
-            insert into CustomRoles (rolename, keycode)
-            values ("Supporter", "Supporter123");
-            insert into CustomRoles (rolename, keycode)
-            values ("Member", "Member123");
-         */
-        for (var i = 0; i < tmp.Count; i++)
-        {
-          for (var j = 0; j < tmp[i].Length; j++)
-          {
-            if (e.Channel.IsPrivate == true && e.Message.Content.StartsWith(tmp[i][1].ToString()))
-            {
-              arrayGuilds = new List<DiscordGuild>(discord.Guilds.Values).ToArray();
-              foreach (var guild in arrayGuilds)
-              {
-                arrayRoles = new List<DiscordRole>(guild.Roles.Values).ToArray();
-
-                foreach (var role in arrayRoles)
-                {
-                  if (role.Name == tmp[i][0].ToString())
-
-                    await globalMember.Member.GrantRoleAsync(role);
-                }
-              }
-            }
-          }
-        }
-
-        
+        await RoleEvents.ReactOnUserMessage(s, e, discord);
       };
       await discord.ConnectAsync();
       await Task.Delay(-1);
